Let all integer operations accept foreign number as second argument
