+++ engine/runtime/src/main/java/org/enso/interpreter/node/expression/builtin/number/integer/AddNode.java
@@ -6,17 +6,13 @@ import com.oracle.truffle.api.dsl.Fallback;
 import com.oracle.truffle.api.dsl.Specialization;
 import com.oracle.truffle.api.interop.InteropLibrary;
 import com.oracle.truffle.api.interop.TruffleObject;
-import com.oracle.truffle.api.interop.UnsupportedMessageException;
 import com.oracle.truffle.api.library.CachedLibrary;
-import com.oracle.truffle.api.nodes.Node.Child;
 import org.enso.interpreter.dsl.BuiltinMethod;
 import org.enso.interpreter.node.expression.builtin.number.utils.BigIntegerOps;
-import org.enso.interpreter.node.expression.builtin.number.utils.ToEnsoNumberNode;
 import org.enso.interpreter.runtime.number.EnsoBigInteger;
 
 @BuiltinMethod(type = "Integer", name = "+", description = "Addition of numbers.")
 public abstract class AddNode extends IntegerNode {
-  private @Child ToEnsoNumberNode toEnsoNumberNode = ToEnsoNumberNode.create();
 
   public abstract Object execute(Object self, Object that);
 
@@ -66,17 +62,7 @@ public abstract class AddNode extends IntegerNode {
       TruffleObject that,
       @CachedLibrary(limit = "3") InteropLibrary iop,
       @Cached AddNode delegate) {
-    try {
-      if (iop.fitsInLong(that)) {
-        return delegate.execute(self, iop.asLong(that));
-      } else if (iop.fitsInDouble(that)) {
-        return delegate.execute(self, iop.asDouble(that));
-      } else if (iop.fitsInBigInteger(that)) {
-        return delegate.execute(self, new EnsoBigInteger(iop.asBigInteger(that)));
-      }
-    } catch (UnsupportedMessageException ex) {
-    }
-    return doOther(self, that);
+    return super.doInterop(self, that, iop, delegate);
   }
 
   @Fallback
