+++ engine/runtime/src/main/java/org/enso/interpreter/node/expression/builtin/number/integer/DivNode.java
@@ -1,19 +1,19 @@
 package org.enso.interpreter.node.expression.builtin.number.integer;
 
+import com.oracle.truffle.api.dsl.Cached;
 import com.oracle.truffle.api.dsl.Fallback;
 import com.oracle.truffle.api.dsl.Specialization;
-import com.oracle.truffle.api.nodes.Node.Child;
+import com.oracle.truffle.api.interop.InteropLibrary;
+import com.oracle.truffle.api.interop.TruffleObject;
+import com.oracle.truffle.api.library.CachedLibrary;
 import org.enso.interpreter.dsl.BuiltinMethod;
 import org.enso.interpreter.node.expression.builtin.number.utils.BigIntegerOps;
-import org.enso.interpreter.node.expression.builtin.number.utils.ToEnsoNumberNode;
 import org.enso.interpreter.runtime.EnsoContext;
 import org.enso.interpreter.runtime.error.DataflowError;
 import org.enso.interpreter.runtime.number.EnsoBigInteger;
 
 @BuiltinMethod(type = "Integer", name = "div", description = "Division of numbers.")
 public abstract class DivNode extends IntegerNode {
-
-  private @Child ToEnsoNumberNode toEnsoNumberNode = ToEnsoNumberNode.create();
 
   abstract Object execute(Object self, Object that);
 
@@ -51,6 +51,15 @@ public abstract class DivNode extends IntegerNode {
   Object doBigInteger(EnsoBigInteger self, EnsoBigInteger that) {
     // No need to trap, as 0 is never represented as an EnsoBigInteger.
     return toEnsoNumberNode.execute(BigIntegerOps.divide(self.getValue(), that.getValue()));
+  }
+
+  @Specialization(guards = "isForeignNumber(iop, that)")
+  Object doInterop(
+      Object self,
+      TruffleObject that,
+      @CachedLibrary(limit = "3") InteropLibrary iop,
+      @Cached DivNode delegate) {
+    return super.doInterop(self, that, iop, delegate);
   }
 
   @Fallback
