+++ engine/runtime/src/main/java/org/enso/interpreter/node/expression/builtin/number/integer/DivideNode.java
@@ -1,14 +1,19 @@
 package org.enso.interpreter.node.expression.builtin.number.integer;
 
+import com.oracle.truffle.api.dsl.Cached;
 import com.oracle.truffle.api.dsl.Fallback;
 import com.oracle.truffle.api.dsl.Specialization;
+import com.oracle.truffle.api.interop.InteropLibrary;
+import com.oracle.truffle.api.interop.TruffleObject;
+import com.oracle.truffle.api.library.CachedLibrary;
 import org.enso.interpreter.dsl.BuiltinMethod;
 import org.enso.interpreter.node.expression.builtin.number.utils.BigIntegerOps;
 import org.enso.interpreter.runtime.number.EnsoBigInteger;
 
 @BuiltinMethod(type = "Integer", name = "/", description = "Division of numbers.")
 public abstract class DivideNode extends IntegerNode {
-  abstract double execute(Object self, Object that);
+  @Override
+  abstract Object execute(Object self, Object that);
 
   static DivideNode build() {
     return DivideNodeGen.create();
@@ -44,8 +49,17 @@ public abstract class DivideNode extends IntegerNode {
     return BigIntegerOps.toDouble(self.getValue()) / that;
   }
 
+  @Specialization(guards = "isForeignNumber(iop, that)")
+  Object doInterop(
+      Object self,
+      TruffleObject that,
+      @CachedLibrary(limit = "3") InteropLibrary iop,
+      @Cached DivideNode delegate) {
+    return super.doInterop(self, that, iop, delegate);
+  }
+
   @Fallback
-  double doOther(Object self, Object that) {
+  Object doOther(Object self, Object that) {
     throw throwTypeErrorIfNotInt(self, that);
   }
 }
