+++ engine/runtime/src/main/java/org/enso/interpreter/node/expression/builtin/number/integer/GreaterOrEqualNode.java
@@ -1,7 +1,11 @@
 package org.enso.interpreter.node.expression.builtin.number.integer;
 
+import com.oracle.truffle.api.dsl.Cached;
 import com.oracle.truffle.api.dsl.Fallback;
 import com.oracle.truffle.api.dsl.Specialization;
+import com.oracle.truffle.api.interop.InteropLibrary;
+import com.oracle.truffle.api.interop.TruffleObject;
+import com.oracle.truffle.api.library.CachedLibrary;
 import org.enso.interpreter.dsl.BuiltinMethod;
 import org.enso.interpreter.node.expression.builtin.number.utils.BigIntegerOps;
 import org.enso.interpreter.runtime.EnsoContext;
@@ -45,6 +49,15 @@ public abstract class GreaterOrEqualNode extends IntegerNode {
   @Specialization
   boolean doBigInteger(EnsoBigInteger self, EnsoBigInteger that) {
     return BigIntegerOps.compare(self.getValue(), that.getValue()) >= 0;
+  }
+
+  @Specialization(guards = "isForeignNumber(iop, that)")
+  Object doInterop(
+      Object self,
+      TruffleObject that,
+      @CachedLibrary(limit = "3") InteropLibrary iop,
+      @Cached GreaterOrEqualNode delegate) {
+    return super.doInterop(self, that, iop, delegate);
   }
 
   @Fallback
